+++ config/bot.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'static_stage_url' => 'https://devstatic.futureskills-sh.de/data/kjdsihadjcfkatbb/bot.json',
+    'static_production_url' => 'https://static.futureskills-sh.de/data/kjdsihadjcfkatbb/bot.json'
+];