+++ database/migrations/2021_01_18_115741_change_offers_table_ext_ids.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class ChangeOffersTableExtIds extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('offers', function (Blueprint $table) {
+            $table->string('ext_id')->nullable();
+            $table->boolean('active')->default(true);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('offers', function($table) {
+            $table->dropColumn('ext_id');
+            $table->dropColumn('active');
+        });
+    }
+}